% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribe.R
\name{tribe}
\alias{tribe}
\alias{tribe<-}
\title{Object attribute list}
\usage{
tribe(obj)

tribe(obj) <- value
}
\arguments{
\item{obj}{An object.}

\item{value}{An appropriate named list of attributes, or \code{NULL}.}
}
\value{
A named list, the attributes of \code{obj}.
}
\description{
The function \code{tribe} is identical to \code{\link[base]{attributes}}, 
expect that it always returns a named list (thus, when \code{attributes} 
will return \code{NULL}, \code{tribe} will return an empty named list).
}
\seealso{
\code{\link[base]{attributes}}, 
\code{\link[base]{attributes<-}}, 
\code{\link[base]{mostattributes<-}}.
}

